Produce report by default and print summary to stdout
Bind tasks so that coverage report is generated by default when `check`
task is run.

Add action to the task that generates clover report that prints a
summary to stdout (using Clover's `clover-log` Ant task). This is so
that the total coverage can be picked up by GitLab CI, and result in
an updated coverage badge.